Add a directive line to the native C program
We definitely want to avoid this complexifying the C layer.

But the configurator needs the ability to instruct the
calling program to either LAUNCH or CANCEL (i.e. not launch).

This new design leaves open the possibility of additional native
directives beyond LAUNCH and CANCEL, although I don't foresee others.

While at it, this commit also improves the native program's exit codes.