Add the beginnings of some unit tests
FileTest.kt contains some simple tests for the File class.
The user home test is probably broken right now, due to the use of ~.
This is intentional, because we want to make this work eventually.

JavaAnalyzerTest.kt will test the Java analyzer, of course.
The .info files contain needed ground truth on the various Java
distros + versions I have looked at so far.