Use function-level @OptIn annotations
At least in my hands, IntelliJ IDEA (2023.3.1 Community) on both Linux
and Windows repeatedly complains about the @OptIn annotation being
missing, even though it is present at the file-wide level on top.
It can be surpressed temporarily by executing the quick fix, which adds
a redundant item to the @OptIn list, then going to the top and deleting
the redundant item. But it's not long before the error comes back.

This change is an attempt to work around IntelliJ's bug here.