Tell Jaunch how many lines of stdin to expect
When passing args from C to Kotlin via stdin, instead of relying on the
stdin pipe to be closed to signal the end of the input, just tell Jaunch
how many lines it is supposed to be reading. Then on the Kotlin side,
read that many lines, and stop, even if stdin is still open.

This seems safer to me than expecting stdin to be able to be immediately
closed on all platforms.