Rework the File class to always canonicalize
Juggling relative vs. absolute paths is a pain, and even an absolute
path isn't enough to know whether a bunch of . and/or .. components are
littered throughout the path expression. Instead of screwing around,
let's just always canonicalize the file paths we are reasoning about.

This commit also implements the ${app-dir} special variable.