Remove especially difficult JavaTest case
The case in question, OpenJDK21U-jdk_x64_linux_hotspot_21.0.1_12,
is not actually the name of any Java installation base directory
I came across. Rather, it is the name of a .tar.gz archive I
downloaded for Eclipse Temurin 21. The actual base directory
inside that archive is jdk-21.0.1+12, which is parseable using
the current logic. So I'm not concerned about complexifying the
path parsing further just to support this theoretical case.