Purge dependencies
The ktoml-based config reader works great. But it is overkill;
Jaunch's config use a tiny subset of the TOML feature set. As such,
the final compiled binary for the Jaunch configurator was much larger
than necessary. This commit rewrites the TOML parsing to do it by hand
without the ktoml library, purging the ktoml and okio dependencies.

This change reduces the size of the compiled binary (at least on Linux)
from ~3.7M to ~1.6M (854K to 381K after compressing with upx).