Really fix the Windows CI build
This works around a weird problem with make + bash, where there is some
sort of mismatch between the MinGW programs and the built-in Windows
ones. It may happen because Jaunch's Makefile targets call shell scripts
that require bash, and somehow the wrong bash is chosen, which triggers
a Windows error about how WSL needs to be installed.

I'm mostly just speculating. But whatever is actually going on, in this
case, using a shebang of `#!/usr/bin/env bash` avoids the issue.

See also actions/runner-images#7253.