Introduce a new dist target
This target builds up a dist/ folder, full of Jaunch build outputs.
It is intended to be zipped up and uploaded as an artifact, as a
starting point for people who want to use Jaunch without building it.

It also runs upx on the binaries under dist/, if Linux or Windows.
(I do not think pre-built upx binaries are available for macOS.)

Meanwhile, the existing app target has been simplified to use the
dist folder as a starting point.