CI: Try to work around download-artifact problems
It seems that the merge-multiple flag is not being respected.
Each artifact is reportedly still being unpacked into its own dir.

Also, as explained at:

    https://github.com/actions/download-artifact#permission-loss

the upload-artifact action does not preserve permissions! I find
this shocking, given that the central use case of this action is
to share built artifacts including native programs. The stated
workaround is to tar up your files first, which would result in
a tar-in-a-zip, which seems strange. So this commit attempts to
roll its own zips; we'll see how it goes.