Fix bug in argument parsing with equals symbol
For example, arguments like -Dfoo=bar were not being parsed correctly;
Jaunch splits to a key/value pair in this case, in case the key appears
in its supported options, which is correct. But in case the key is *not*
on Jaunch's supported options list, the whole argument should be used,
not only the key.