Fix more argument handling bugs
Recognized JVM arguments all start with a dash. My original plan was to
make such dashes implicit, with the the Jaunch configurator adding
them... but upon reflection I think it's better for the TOML file to be
explicit. Otherwise, there is no way to properly support a JVM that
offers any non-dash-prefixed arguments.