Add a built-in ${executable} variable
It matches the executable passed as the first argument to Jaunch.
The behavior implemented in this commit matches the old ImageJ Launcher
behavior... but I am not sure it's actually very useful, because Fiji
changes the CWD to the application directory, at which point the
relative path given to execute the native launcher is no longer correct.
Maybe better would be to set it to an absolute path always?