Cross-compile arm64 targets for Linux and macOS
For compiling the native launcher portion of Jaunch on Linux,
this requires the gcc-aarch64-linux-gnu package to be installed,
for the aarch64-linux-gnu-gcc command line tool.

For the configurator portion of Jaunch, Gradle should take care
of bootstrapping the necessary compilation tools.

I had a heck of a time figuring out the correct Gradle ceremony,
even with the help of Internet searches and ChatGPT, so I posted
an answer on StackOverflow about it to hopefully help others:

    https://stackoverflow.com/a/77964683/1207769