Make the configurator support multiple directives
This is the other half of the work done in
b09a7e11c0619606c83cf2f125c42426e745dcf2.

This patch is kind of a monster (sorry about that), because it
restructures the previously very long main function into a series
of helper functions, to make it easier to understand and navigate.