Remove resolved TODO
The R_OK flag tests for read access, whereas F_OK tests for existence.
So F_OK is correct here, for the File API. If we also need to test for
read (and/or write and/or execute) access at some point, we can enable
the commented out canRead/Write/Execute functions as well at that time.