Improve runtime-specific directive handling
When using --dry-run, we don't want all enabled runtimes to handle it,
only *activated* runtimes -- i.e., the runtimes that would be launched.

But if no activated runtimes *wants* to handle it, then as a fallback,
we can ask the other enabled (but dormant) runtimes to take care of it.

Not sure yet whether this is a good heuristic in all cases, but I think
it will do the right thing for the current configurator-side directives
--dry-run, --print-(python|java)-home, and --print-(python|java)-info.