Add update logic to configurator
This assumes there is an "update" app subdirectory with a directory
structure shadowing that of the main app. Empty update files signify the
corresponding app file should be deleted; otherwise files are copied
from the update structure to their equivalent app location. The update
directory and any remaining contets are deleted.

Co-authored-by: Curtis Rueden <ctrueden@wisc.edu>