Add Windows console wrangling logic
This logic successfully attaches to the parent process's existing
console if any, although stdin is a mess because unlike Windows console
applications, Windows GUI applications return control to the shell
immediately rather than blocking until the invoked program completes
execution, which results in each typed character being sent arbitrarily
to either the Jaunched app or to the shell; "there is no guarantee that
input is received by the process for which it was intended"
(https://learn.microsoft.com/en-us/windows/console/creation-of-a-console).

Still, this behavior is better than no console at all.