Enable 2-stage runtime parameter processing
Previously we were performing special-case parameter processing (e.g.
expanding %'s in jvmMaxHeap specifications) at the time of initial
runtime configuration. However, this would happen a) before variable
interpolation and b) only if the user hadn't manually specified a value.
As a result, %'s would only work in a very limited number of cases (when
set in the .toml)

Instead, we now do a second "processArgs" pass that happens after
variable interpolation. It is intended to operate on all runtime args at
that point, whether they came from a toml configuration, user, etc...

This allows the particular case of % calculation in jvmMaxHeap to
trigger in all cases.

Closes #64.

Note that because we process every item in the arg list we are working
around #63, without actually fixing it.