Log debug output to file matching executable name
We do this by remembering the config file we gleaned during our scan
for TOMLs, and assigning the log file path to match what we found...
unless the JAUNCH_LOGFILE environment variable is set, in which case
it takes precedence.

Our mission is complicated by the fact that we don't actually know the
name of the matching config file until we are a few log statements into
the configuration process. So in case a line is logged before the log
file path is known, we save it to a buffer, and then flush it out later
once the log file path is known.