Ensure hints match the var key for aliased fields
Jaunch config fields are entered as hints to trigger the appropriate
actions. For assignment fields, there is a corresponding K:V entry in
the vars.

Config fields can have multiple key aliases, but regardless of which is
entered we always add the first "primary" alias to the vars table.

We must also register the hint matching this vars entry, otherwise the
vars entry will be ignored when processing the hint and the intended
function will fail. With this change, this now happens correctly.

Previously we were recording the hint using the actual alias, resulting
in aliases not working.

Closes https://github.com/apposed/jaunch/issues/62