Emit configurator failure message into the log too
In debug mode, everything at the configurator level is supposed to be
logged to a log file. But the when the configuration fails, the
configurator was only reporting the error back to the calling native
launcher process, and not passing it through the logging mechanism.
So the log file would not show the error message.

Furthermore, in case the configuration fails so early that the log file
has not yet been created (because we don't yet know the ideal filename
for it), we should still flush out the pending log lines to *some* log
file before failing out; in want of a better name, we use `jaunch.log`.