Fix configurator subprocess stderr handling
We now successfully see all output in debug mode in all of PowerShell,
Command Prompt, and Git Bash scenarios. There are still some badly
behaved edge cases, such as a bash shell invoked from inside CMD or PS,
but what do you want, I'm not a miracle worker!

We also now emit a warning if launching in a potentially sideways
scenario (CMD or PS), including hints on how to do it more successfully.
Unfortunately, this logic doesn't detect whether any of the workarounds
were used, particularly whether a batch file was used, which would be
nice to know so that we don't warn spuriously. Maybe later...