Fix ABORT directive protocol mismatch
I waffled on whether the ABORT directive should print an output line
count of 0, to keep it consistent with other no-args directives (e.g.
INIT_THREADS) that do so. But as of the previous commit, the
configurator was emitting a 0 while the launcher was expected no 0.

This commits reconciles the behavior, removing the emission of 0.
If we need arguments to ABORT in the future, we can update both the
launcher and configurator simultaneously to accommodate that.

This should fix the confusing "Ignoring 1 trailing output lines"
message whenever application launch gets aborted.