Revise how executable packing logic works
This commit changes the shell script to be standalone, accepting an
explicit list of executable files as arguments, and removes the pack
goal from the build system.

Already we've had reports of Windows anti-malware misidentifying
the Jaunch configurator as infected by TROJ.Win32.TRX.XXPE50FLM011.
As such, let's make the packing step optional, so that it can be
invoked on a per-binary basis, and avoided in cases where it
triggers false positives from security scanners.