Fix handling of ${appName}.app/Contents/MacOS
When launching from inside a macOS bundled app, we want to set the
application directory to the directory *outside* that app bundle,
because these days on macOS, app bundles need to be code-signed and
immutable. In other words: the .app is merely the minimal launcher;
the actual application codebase lives *outside* the bundled app.