Fix problem with Windows code signing script
The script was erroring out at the signtool.exe step with:

> SignTool Error: No file digest algorithm specified. Please specify the
> digest algorithm with the /fd flag. Using /fd SHA256 is recommended
> and more secure than SHA1. Calling signtool with /fd sha1 is
> equivalent to the previous behavior. In order to select the hash
> algorithm used in the signing certificate's signature, use the /fd
> certHash option.

The issue was that MSYS2 was doing path conversion on the terms with
slashes, replacing the forward slashes with backslashes in an overeager
attempt to make Windows happy. But we need these arguments passed as-is,
and Windows these days mostly understands paths with forward slashes
(signtool.exe certainly does), so we can just suppress that behavior.