Recommend using OS-specific icon files in SETUP
The blanket --app-icon option still exists. But it is not convenient
enough to use, because: A) you must have additional utilities installed
like png2icns and magick; and B) icons on macOS are supposed to look
different than they do on other platforms, with a rounded rectangle
frame. Jaunch could get into the business of overlaying the given SVG
icon onto such a frame, but choices must be made about background color
and such, which are probably better left to the application developer.

Even breaking apart the --app-icon option into three options in the
guide is still not great though, because Microsoft's signtool.exe fails
to re-sign binaries that were already signed and subsequently added an
icon. The solution is to first un-sign the binary, then inject the icon,
and then re-sign it. But we can only do that on Windows systems. So as
things stand, appify.sh won't do the job when run on Linux or macOS. :(