Support relative jvmRootPaths outside CWD
If we have a relative path (from the app directory), e.g. from a config
file, we need to at some point resolve it relative to the app-dir. This
was not being done previously so relative dirs only worked when
CWD=app-dir.

With this fix we add a pre-filtering pass where we try to resolve
non-directory JVM root paths as relative to app-dir.