Let --jaunch-debug enable Jaunch's debug mode
There is an important distinction here: the `--debug` flag enables
Jaunch's debug mode *but also* gets passed along to the rest of the
argument parsing infrastructure, so that e.g. it can be handed off
to the main program, as configured within the TOML configuration.

In contrast, `--jaunch-debug` is a purely internal Jaunch flag, which
will never be visible downstream from Jaunch's initial argument parsing.