Appify faster if JAUNCH_APPIFY_FASTER var is set
Specifically, the ImageMagick and Wine invocations have a high cost.
But these are only needed to embed icons, which is not necessary
for running the integration tests. When constructing the demos
locally for the purpose of running those tests, it's nice for it to
happen more quickly. So say hello to this hacky flag!