Fix bugs in OS & ARCH canonicalization
It is too dangerous to attempt canonicalization of these internal flag
values while configuration TOML files are still in the process of being
merged. Better to save the internal flags as is into the configuration
(more consistent with the saving of raw values from the TOML otherwise),
and then offer smart accessors that do just-in-time canonicalization on
the finalized config object, to ensure we utilize the broadest available
set of declared osAliases and archAliases.