Let directives specify an argument within the TOML
When a directive is followed by a colon, the string afterward is now
emitted as a single line (the first line) of the directive's arguments.

The current intended use cases for this feature are with RUNLOOP and
SETCWD, each of which needs a single argument following the directive
itself.

I chose colon because:

1. File paths rarely contain the colon symbol.
   - Linux: Only illegal character in filenames.
   - macOS: Historically avoided, now illegal in most contexts.
   - Windows: Reserved for drive letters (C:), not in directory names.

2. Consistency with existing patterns
   - Many languages and config formats use : for key-value pairs
   - Shell PATH uses : as separator on POSIX-friendly platforms

I also considered:

* Semicolon (;) -- But elements of the directives list are already
  split on comma (,) as a separator between directives, and semicolons
  are generally considered a wider separator than commas, so their
  use here would invert that assumption (violating PoLA).

* Space ( ) -- But many file paths have spaces, and my feeling is that
  committing to spaces here will make it more complex later to add
  multi-argument TOML-based directive arguments (COPY srcFile destFile),
  because it would immediately require an escaping mechanism for spaces.

* TOML lists or inline tables -- But use of that more complex TOML
  syntax would make the configurator's TOML parsing code more involved.

So colon seems the least problematic choice. But even with colons,
adding support for multiple TOML-specified directive arguments would
probably require an new escaping mechanism due to Windows drive letters.
To avoid that extra work for the moment, this commit only adds support
for a single TOML-specified directive argument per directive.