Launch Python better
This commit changes the Python launch to pass the Python *executable*
path to Py_BytesMain rather than passing the Python *library* path as
was being done previously. This is especially important on Windows,
where passing the library path does not work in all scenarios (see #87).

We accomplish this by first looking for the Python executable launcher,
and then invoking it to find the associated Python shared library.
That is: Jaunch runs a special Python script (configs/props.py) to glean
the location of libpython. This approach is much more robust across the
variety of platforms and environment flavors out there, including
conda/mamba, uv, Homebrew, and system Python installations, with
flexibility to adapt to more kinds of environments as they emerge.
(I did not test with pixi yet, but hopefully it works as well.)

Unfortunately, implementing this approach led to breaking changes:

* Jaunch's PYTHON directive now requires an extra line -- the path to
  the Python executable -- in addition to the other lines already given.

* The TOML key python.lib-suffixes became python.exe-suffixes instead.
  This is actually a nice change, as it simplifies that key's default
  rules, but it is nonetheless a breaking change to the schema.

Therefore, this commit updates the Jaunch dev version to 2.0.0-SNAPSHOT.

Closes #87.