Make the integration tests work on macOS again
After the "follow symlinks" fix, Jaunch's behavior on macOS changed.
Because the demo/hi-macos* executables are symlinks to
Hi.app/Contents/MacOS/hi-macos*, and the executable is now resolved to
the latter, the configurator at Hi.app/Contents/MacOS/jaunch-macos* is
now selected, rather than the copy at jaunch/jaunch*, as previously.
So the sequence.t's approach of replacing jaunch/jaunch* was no longer
sufficient to alter the configurator behavior on macOS.