Constrain Python installations by target OS + arch
The Java installations are already so constrained.

You might think: Python should already be constrained, because we
wouldn't be able to invoke `python props.py` if the Python installation
in question wasn't the correct OS and architecture! But there is a
notable edge case: Windows arm64, which supports execution of Windows
x64 binaries via the Prism compatibility layer. So it's particularly
important to do this filtering when launching on Windows arm64.