Loading Windows DLLs even harder
With 65b7a9ff, Jaunch began adding the to-be-loaded DLL's directory to
the DLL search path using SetDllDirectory. This worked in a pinch for
Python, but it is not good enough for Java on some Windows machines.

Claude Code and I played with using AddDllDirectory plus LoadLibraryEx,
but that also is insufficient: jvm.dll is then loaded successfully, but
after the Java program starts up, as soon as more DLL loading needs to
happen -- such as to initialize AWT -- we hit another DLL path failure
because because after loading jvm.dll, some other dependent Java DLL
(e.g. awt.dll) can then not (re?)load jvm.dll using plain LoadLibrary.

So ultimately, we settled on the classic Windows mainstay: adding
directories to the PATH environment variable. And just for Java, we
use an ugly hack: if the DLL to be loaded lies within a subdirectory of
bin, add not only the DLL directory, but also its parent bin directory,
to the PATH; this is necessary because OpenJDK installations typically
put jvm.dll into bin\server and the rest of the DLLs directly into bin.