Categorize directives by pre-colon portion only
The Jaunch configurator classifies a directive as a launch directive
iff the directive is in ALL-CAPS. But for directives with an argument
like SETCWD:/path/to/something, the argument portion following the
colon should not affect the classification. The fix is to check only
the portion of the directive string before the colon.

Fixes #96.